+++ event-sourcing/event-publisher/event-publisher-timer/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/prepublish/PrePublisherTimerConfig.java
@@ -1,13 +1,17 @@
 package uk.gov.justice.services.eventsourcing.publishedevent.prepublish;
 
 import static java.lang.Boolean.parseBoolean;
+import static java.lang.Double.parseDouble;
 import static java.lang.Long.parseLong;
 
 import uk.gov.justice.services.common.configuration.GlobalValue;
+import uk.gov.justice.services.common.configuration.Value;
+import uk.gov.justice.services.eventsourcing.publishedevent.NotifierWorkerConfig;
 
+import javax.annotation.PostConstruct;
 import javax.inject.Inject;
 
-public class PrePublisherTimerConfig {
+public class PrePublisherTimerConfig extends NotifierWorkerConfig {
 
     @Inject
     @GlobalValue(key = "pre.publish.start.wait.milliseconds", defaultValue = "7250")
@@ -24,6 +28,22 @@ public class PrePublisherTimerConfig {
     @Inject
     @GlobalValue(key = "pre.publish.disable", defaultValue = "false")
     private String disablePrePublish;
+
+    @Inject
+    @Value(key = "event.linking.worker.notified", defaultValue = "false")
+    private String eventLinkerNotified;
+
+    @Inject
+    @Value(key = "event.linking.worker.backoff.min.milliseconds", defaultValue = "5")
+    private String backoffMinMilliseconds;
+
+    @Inject
+    @Value(key = "event.linking.worker.backoff.max.milliseconds", defaultValue = "100")
+    private String backoffMaxMilliseconds;
+
+    @Inject
+    @Value(key = "event.linking.worker.backoff.multiplier", defaultValue = "1.5")
+    private String backoffMultiplier;
 
     public long getTimerStartWaitMilliseconds() {
         return parseLong(timerStartWaitMilliseconds);
@@ -43,5 +63,13 @@ public class PrePublisherTimerConfig {
 
     public void setDisabled(final boolean disable) {
         this.disablePrePublish = Boolean.toString(disable);
+    }
+
+    @PostConstruct
+    public void postConstruct() {
+        this.setShouldWorkerNotified(parseBoolean(eventLinkerNotified));
+        this.setBackoffMinMilliseconds(parseLong(backoffMinMilliseconds));
+        this.setBackoffMaxMilliseconds(parseLong(backoffMaxMilliseconds));
+        this.setBackoffMultiplier(parseDouble(backoffMultiplier));
     }
 }